+++ layout/index.ts
@@ -1,3 +1,5 @@
+export * from './DocumentContext'
+export * from './NavContext'
 export * from './MDXLayout'
 export * from './MDXLayoutNav'
 export * from './MDXLayoutPagination'
